requirements.txt
langchain
langchain-openai
langchain-community
pypdf
python-dotenv
faiss-cpu